Add two methods on adjMatrix class
I added two operations on the adjacencyMatrix.cpp file. They are
getAdjacentNodes, that returns a vector<int> that contains the indexes
of all the vertices that are adjacent to the vertex passed as a
parameter of the function, and areAdjacent, that returns a boolean that
signals if any two vertices are adjacent.