Complete removeEdge function
This commits adds another if to the removeEdge function to deal with the
possiblity that we are removing an edge from an undirecte graph. That
is, we need to remove an element from two lists, not just one,
representing that neither of the vertices are connected to each other
anymore.